+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Logrus Graylog hook
 
+## 2.0.1 - 2016-08-16
+
+* Fix an issue with entry constructor (#12)
+
 ## 2.0.0 - 2016-07-02
 
 * Remove facility param in constructor, as it's an optional param in Graylog 2.0 (credits: @saward / #9)